Refactor main reducer logic
The problem with the previous approach was that the actions sent to the
reducer were not specific enough and there was a lot of logic in the
reducer that was only there to interpret what the "real" state update
should be.

For example, if the reducer was passed a PLAY action, it had to check if
state.uiState === PAUSED, and if so, instead of treating the action like
a true PLAY action and updating state.game it had to clear the interface
pause state and *not* play a card, which was very unintuitive.

Likewise a CONTINUE action might actually have been a no-op if there was
no pause state.

To fix this the actions were redefined to be more specific. Now PLAY
is only used when the user has clicked on a card, and a new action
AI_MOVE is used for computer turns.

Additionally, a new function actionFromClick was created to convert a
canvas click event into the correct action.